Use argparse to provide optional connection string.

+++ misc/e2etest.py
@@ -121,6 +121,7 @@ def main():
     ap = argparse.ArgumentParser()
     ap.add_argument('--keep-temps', default=False, action='store_true')
     ap.add_argument('--verbose', default=False, action='store_true')
+    ap.add_argument('--connection-string', help='Use this connection string instead of attempting to manage a local database.')
     args = ap.parse_args()
     if args.verbose:
         logging.basicConfig(level=logging.DEBUG)
@@ -136,14 +137,19 @@ def main():
             logger.info("leaving temp dir %r", tdir.name)
     ensureTestData(e2edata)
 
-    dbname = 'e2eindex_{}_{}'.format(int(time.time()), random.randrange(1000))
-    xrun(['dropdb', '--if-exists', dbname], timeout=5)
-    xrun(['createdb', dbname], timeout=5)
-    if not args.keep_temps:
-        atexitrun(['dropdb', '--if-exists', dbname], timeout=5)
+    # Setup database connection string.
+    if args.connection_string is None:
+        dbname = 'e2eindex_{}_{}'.format(int(time.time()), random.randrange(1000))
+        xrun(['dropdb', '--if-exists', dbname], timeout=5)
+        xrun(['createdb', dbname], timeout=5)
+        if not args.keep_temps:
+            atexitrun(['dropdb', '--if-exists', dbname], timeout=5)
+        else:
+            logger.info("leaving db %r", dbname)
+        psqlstring = 'dbname={} sslmode=disable'.format(dbname)
     else:
-        logger.info("leaving db %r", dbname)
-    psqlstring = 'dbname={} sslmode=disable'.format(dbname)
+        psqlstring = args.connection_string
+
     xrun(['cmd/algorand-indexer/algorand-indexer', 'import', '-P', psqlstring, os.path.join(e2edata, 'blocktars', '*'), '--genesis', os.path.join(e2edata, 'algod', 'genesis.json')], timeout=20)
     cmd = ['cmd/algorand-indexer/algorand-indexer', 'daemon', '-P', psqlstring, '--dev-mode', '--no-algod']
     logger.debug("%s", ' '.join(map(repr,cmd)))
